--- src/main/java/gibson/Launcher.java
@@ -1,9 +1,0 @@
-package gibson;
-
-import javafx.application.Application;
-
-public class Launcher {
-    public static void main(String[] args) {
-        Application.launch(Ui.class, args);
-    }
-}
